Server/image_functions/requirements.txt
scikit-image
numpy
pillow
matplotlib
pybase64